+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed a PHP error that could occur when viewing variant indexes and using Craft CMS 5.6.17 or earlier. ([#4060](https://github.com/craftcms/commerce/issues/4060))
 - Fixed a bug where currency based order condition rules incorrectly showed up as custom fields.
 - Fixed a SQL error that could occur when restoring Variants that were previously soft-deleted. ([#4065](https://github.com/craftcms/commerce/issues/4065))
+- Fixed a bug where Subscription and Transfer elements’ field layouts were not deleted when Commerce was uninstalled.
 
 ## 5.4.0 - 2025-06-23